+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ChangeLog
 0.0.17 (2026-05-02)
 -------------------
 
+- IMPROVEMENT: ensure tools like "npm" and "claude" are found in $PATH
 - CLEANUP: update documentation for usage of plugin and tool
 
 0.0.16 (2026-05-02)